src/api/user/serializers/user_public.py
from rest_framework import serializers

from src.apps.user.models import User


class UserPublicSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'email', 'first_name')
